Setting up project cards page.
