+++ src/App.js
@@ -4,6 +4,7 @@ import Navigation from "./components/Navigation";
 import Footer from "./components/Footer";
 import HomePage from "./pages/HomePage";
 import AboutMePage from "./pages/AboutMePage";
+import ProjectCardsPage from "./pages/ProjectCardsPage";
 // import Skills from "./components/Skills";
 // import Projects from "./components/Projects";
 // import Contact from "./components/Contact";
@@ -17,7 +18,8 @@ function App() {
         <Navigation />
         <Routes>
           <Route path="/" element={<HomePage />}></Route>
-          <Route path="/about-me-page" element={<AboutMePage />}></Route>
+          <Route path="/about-me" element={<AboutMePage />}></Route>
+          <Route path="/project-cards" element={<ProjectCardsPage />}></Route>
 
           {/* <Route path="/React-Portfolio" element={<About />}></Route>
           <Route path="/skills" element={<Skills />}></Route>
